+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   * new: added resolution hints for some known issues.
   * new: extended value expression `@extend:` finds and replace extended values in a string (e.g. for JSON)
   * new: orchestrator: error analysis for workflow start
+  * new: node: added command `transfer sessions` to list all sessions of all transfers
 
 * 4.14.0